Made the code more robust when the SD card fails
If the SD libraray does not initialize, then status is recorded and
not cleared. The SD code does not get run. The SD card
functions called from setup() do not test the status flag.
Also, the code that wakes up the node must not be run if SD
initialization fails. This seems to affect the LORA.